chore: `InitiativeDetailsBaseScreenComponent`
