+++ ts/features/idpay/initiative/details/screens/InitiativeDetailsScreen.tsx
@@ -1,39 +1,31 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
 import { useNavigation, useRoute } from "@react-navigation/core";
 import { RouteProp, useFocusEffect } from "@react-navigation/native";
-import { format } from "date-fns";
+import { sequenceS } from "fp-ts/lib/Apply";
 import * as O from "fp-ts/lib/Option";
 import { pipe } from "fp-ts/lib/function";
-import React, { useCallback } from "react";
-import { ScrollView, StyleSheet, View } from "react-native";
-import Placeholder from "rn-placeholder";
+import * as React from "react";
+import { StyleSheet, View } from "react-native";
 import {
   InitiativeDTO,
-  InitiativeRewardTypeEnum
+  InitiativeRewardTypeEnum,
+  StatusEnum as InitiativeStatusEnum
 } from "../../../../../../definitions/idpay/InitiativeDTO";
 import { ContentWrapper } from "../../../../../components/core/ContentWrapper";
 import { Pictogram } from "../../../../../components/core/pictograms";
 import { VSpacer } from "../../../../../components/core/spacer/Spacer";
 import { Body } from "../../../../../components/core/typography/Body";
 import { H3 } from "../../../../../components/core/typography/H3";
-import { LabelSmall } from "../../../../../components/core/typography/LabelSmall";
-import { IOColors } from "../../../../../components/core/variables/IOColors";
-import { IOStyles } from "../../../../../components/core/variables/IOStyles";
-import BaseScreenComponent from "../../../../../components/screens/BaseScreenComponent";
 import ButtonSolid from "../../../../../components/ui/ButtonSolid";
-import FocusAwareStatusBar from "../../../../../components/ui/FocusAwareStatusBar";
 import I18n from "../../../../../i18n";
 import {
   AppParamsList,
   IOStackNavigationProp
 } from "../../../../../navigation/params/AppParamsList";
 import { useIODispatch, useIOSelector } from "../../../../../store/hooks";
-import { emptyContextualHelp } from "../../../../../utils/emptyContextualHelp";
 import { IDPayConfigurationRoutes } from "../../configuration/navigation/navigator";
-import {
-  InitiativeCardComponent,
-  InitiativeCardComponentSkeleton
-} from "../components/InitiativeCardComponent";
+import { InitiativeBonusCounter } from "../components/InitiativeBonusCounter";
+import InitiativeDetailsBaseScreenComponent from "../components/InitiativeDetailsBaseScreenComponent";
 import { InitiativeSettingsComponent } from "../components/InitiativeSettingsComponent";
 import InitiativeTimelineComponent from "../components/InitiativeTimelineComponent";
 import { MissingConfigurationAlert } from "../components/MissingConfigurationAlert";
@@ -62,10 +54,18 @@ const InitiativeDetailsScreen = () => {
   const dispatch = useIODispatch();
   const initiativeDataPot = useIOSelector(idpayInitiativeDetailsSelector);
 
-  const initiativeData: InitiativeDTO | undefined = pot.getOrElse(
-    initiativeDataPot,
-    undefined
-  );
+  const navigateToBeneficiaryDetails = () => {
+    navigation.navigate(IDPayDetailsRoutes.IDPAY_DETAILS_MAIN, {
+      screen: IDPayDetailsRoutes.IDPAY_DETAILS_BENEFICIARY,
+      params: {
+        initiativeId,
+        initiativeName: pot.getOrElse(
+          pot.map(initiativeDataPot, initiative => initiative.initiativeName),
+          undefined
+        )
+      }
+    });
+  };
 
   const navigateToConfiguration = () => {
     navigation.navigate(IDPayConfigurationRoutes.IDPAY_CONFIGURATION_MAIN, {
@@ -75,11 +75,12 @@ const InitiativeDetailsScreen = () => {
   };
 
   const navigateToPaymentAuthorization = () => {
-    alert("hello");
+    // TODO payment authorization
+    alert("TBD");
   };
 
   useFocusEffect(
-    useCallback(() => {
+    React.useCallback(() => {
       dispatch(idpayInitiativeGet.request({ initiativeId }));
       dispatch(idpayTimelinePageGet.request({ initiativeId, page: 0 }));
     }, [dispatch, initiativeId])
@@ -89,185 +90,224 @@ const InitiativeDetailsScreen = () => {
     initiativeNeedsConfigurationSelector
   );
 
-  const lastUpdateComponent = pipe(
-    initiativeData?.lastCounterUpdate,
-    O.fromNullable,
-    O.map(date => format(date, "DD/MM/YYYY, HH:mm")),
-    O.fold(
-      () => (
-        <View style={styles.lastUpdate}>
-          <Placeholder.Box animate="fade" width={180} height={16} radius={4} />
-        </View>
-      ),
-      lastUpdated => (
-        <LabelSmall style={styles.lastUpdate} color="bluegrey" weight="Regular">
-          {I18n.t(
-            "idpay.initiative.details.initiativeDetailsScreen.configured.lastUpdated"
-          )}
-          {lastUpdated}
-        </LabelSmall>
+  const getInitiativeCounters = (
+    initiative: InitiativeDTO
+  ): ReadonlyArray<InitiativeBonusCounter> | undefined => {
+    const isRefundable = initiative.status === InitiativeStatusEnum.REFUNDABLE;
+
+    const availableAmount = initiative.amount || 0;
+
+    const amountProgress = pipe(
+      sequenceS(O.Monad)({
+        amount: O.fromNullable(initiative.amount),
+        accrued: O.fromNullable(initiative.accrued)
+      }),
+      O.map(({ amount, accrued }) => ({ total: amount + accrued, amount })),
+      O.filter(({ total }) => total !== 0),
+      O.map(({ amount, total }) => (amount / total) * 100.0),
+      O.getOrElse(() => 100.0)
+    );
+
+    const refundableAmount = pipe(
+      sequenceS(O.Monad)({
+        accrued: O.fromNullable(initiative.accrued),
+        refunded: O.fromNullable(initiative.refunded)
+      }),
+      O.map(({ accrued, refunded }) => accrued - refunded),
+      O.getOrElse(() => 0)
+    );
+
+    return pipe(
+      initiative.initiativeRewardType,
+      O.fromNullable,
+      O.alt(() => O.some(InitiativeRewardTypeEnum.REFUND)),
+      O.fold(
+        () => undefined,
+        (type): ReadonlyArray<InitiativeBonusCounter> => {
+          switch (type) {
+            case InitiativeRewardTypeEnum.DISCOUNT:
+              return [
+                {
+                  type: "AmountWithProgress",
+                  label: I18n.t(
+                    "idpay.initiative.details.initiativeCard.availableAmount"
+                  ),
+                  amount: availableAmount,
+                  progress: amountProgress,
+                  isDisabled: !isRefundable
+                }
+              ];
+            case InitiativeRewardTypeEnum.REFUND:
+              return [
+                {
+                  type: "AmountWithProgress",
+                  label: I18n.t(
+                    "idpay.initiative.details.initiativeCard.availableAmount"
+                  ),
+                  amount: availableAmount,
+                  progress: amountProgress,
+                  isDisabled: !isRefundable
+                },
+                {
+                  type: "Amount",
+                  label: I18n.t(
+                    "idpay.initiative.details.initiativeCard.toRefund"
+                  ),
+                  amount: refundableAmount,
+                  isDisabled: !isRefundable
+                }
+              ];
+          }
+        }
       )
-    )
-  );
-
-  const missingConfigurationAlertComponent = pipe(
-    initiativeData,
-    O.fromNullable,
-    O.map(({ status }) => status),
-    O.fold(
-      () => null,
-      status => (
-        <MissingConfigurationAlert
-          initiativeId={initiativeId}
-          status={status}
-        />
-      )
-    )
-  );
-
-  const renderScreenContent = () => {
-    if (
-      initiativeData?.initiativeRewardType === InitiativeRewardTypeEnum.DISCOUNT
-    ) {
-      return (
-        <ContentWrapper>
-          {lastUpdateComponent}
-          {missingConfigurationAlertComponent}
-          <VSpacer size={8} />
-          <InitiativeTimelineComponent initiativeId={initiativeId} />
-          <VSpacer size={32} />
-        </ContentWrapper>
-      );
-    }
-
-    if (initiativeNeedsConfiguration) {
-      return (
-        <View style={styles.newInitiativeMessageContainer}>
-          <Pictogram name="setup" size={72} />
-          <VSpacer size={16} />
-          <H3>
-            {I18n.t(
-              "idpay.initiative.details.initiativeDetailsScreen.notConfigured.header"
-            )}
-          </H3>
-          <VSpacer size={8} />
-          <Body style={{ textAlign: "center" }}>
-            {I18n.t(
-              "idpay.initiative.details.initiativeDetailsScreen.notConfigured.footer",
-              { initiative: initiativeData?.initiativeName }
-            )}
-          </Body>
-          <VSpacer size={16} />
-          <ButtonSolid
-            accessibilityLabel={I18n.t(
-              "idpay.initiative.details.initiativeDetailsScreen.configured.startConfigurationCTA"
-            )}
-            fullWidth={true}
-            color="primary"
-            onPress={navigateToConfiguration}
-            label={I18n.t(
-              "idpay.initiative.details.initiativeDetailsScreen.configured.startConfigurationCTA"
-            )}
-          />
-        </View>
-      );
-    }
-
-    return (
-      <ContentWrapper>
-        {lastUpdateComponent}
-        {missingConfigurationAlertComponent}
-        <VSpacer size={8} />
-        <InitiativeTimelineComponent initiativeId={initiativeId} />
-        <VSpacer size={24} />
-        <InitiativeSettingsComponent initiative={initiativeData} />
-        <VSpacer size={32} />
-      </ContentWrapper>
     );
   };
 
-  const handleBeneficiaryDetailsPress = () => {
-    navigation.navigate(IDPayDetailsRoutes.IDPAY_DETAILS_MAIN, {
-      screen: IDPayDetailsRoutes.IDPAY_DETAILS_BENEFICIARY,
-      params: {
-        initiativeId,
-        initiativeName: initiativeData?.initiativeName
-      }
-    });
-  };
+  const getInitiativeDetailsContent = (initiative: InitiativeDTO) =>
+    pipe(
+      initiative.initiativeRewardType,
+      O.fromNullable,
+      O.alt(() => O.some(InitiativeRewardTypeEnum.REFUND)),
+      O.fold(
+        () => undefined,
+        rewardType => {
+          switch (rewardType) {
+            case InitiativeRewardTypeEnum.DISCOUNT:
+              return (
+                <ContentWrapper>
+                  <VSpacer size={8} />
+                  <InitiativeTimelineComponent initiativeId={initiativeId} />
+                  <VSpacer size={32} />
+                </ContentWrapper>
+              );
 
-  const cardComponent = pipe(
-    initiativeData,
-    O.fromNullable,
+            case InitiativeRewardTypeEnum.REFUND:
+              if (initiativeNeedsConfiguration) {
+                return (
+                  <View style={styles.newInitiativeMessageContainer}>
+                    <Pictogram name="setup" size={72} />
+                    <VSpacer size={16} />
+                    <H3>
+                      {I18n.t(
+                        "idpay.initiative.details.initiativeDetailsScreen.notConfigured.header"
+                      )}
+                    </H3>
+                    <VSpacer size={8} />
+                    <Body style={{ textAlign: "center" }}>
+                      {I18n.t(
+                        "idpay.initiative.details.initiativeDetailsScreen.notConfigured.footer",
+                        { initiative: initiative.initiativeName }
+                      )}
+                    </Body>
+                    <VSpacer size={16} />
+                    <ButtonSolid
+                      accessibilityLabel={I18n.t(
+                        "idpay.initiative.details.initiativeDetailsScreen.configured.startConfigurationCTA"
+                      )}
+                      fullWidth={true}
+                      color="primary"
+                      onPress={navigateToConfiguration}
+                      label={I18n.t(
+                        "idpay.initiative.details.initiativeDetailsScreen.configured.startConfigurationCTA"
+                      )}
+                    />
+                  </View>
+                );
+              }
+
+              return (
+                <ContentWrapper>
+                  <MissingConfigurationAlert
+                    initiativeId={initiativeId}
+                    status={initiative.status}
+                  />
+                  <VSpacer size={8} />
+                  <InitiativeTimelineComponent initiativeId={initiativeId} />
+                  <VSpacer size={24} />
+                  <InitiativeSettingsComponent initiative={initiative} />
+                  <VSpacer size={32} />
+                </ContentWrapper>
+              );
+          }
+        }
+      )
+    );
+
+  const getInitiativeFooterCta = (initiative: InitiativeDTO) =>
+    pipe(
+      initiative.initiativeRewardType,
+      O.fromNullable,
+      O.alt(() => O.some(InitiativeRewardTypeEnum.REFUND)),
+      O.fold(
+        () => undefined,
+        rewardType => {
+          switch (rewardType) {
+            case InitiativeRewardTypeEnum.DISCOUNT:
+              return (
+                <>
+                  <ButtonSolid
+                    label="Autorizza pagamento"
+                    accessibilityLabel="Autorizza pagamento"
+                    icon="qrCode"
+                    onPress={navigateToPaymentAuthorization}
+                    fullWidth={true}
+                  />
+                  <VSpacer size={16} />
+                </>
+              );
+
+            case InitiativeRewardTypeEnum.REFUND:
+              return undefined;
+          }
+        }
+      )
+    );
+
+  return pipe(
+    initiativeDataPot,
+    pot.toOption,
     O.fold(
-      () => <InitiativeCardComponentSkeleton />,
-      initiative => <InitiativeCardComponent initiative={initiative} />
+      () => (
+        <InitiativeDetailsBaseScreenComponent
+          isLoading={true}
+          onHeaderDetailsPress={navigateToBeneficiaryDetails}
+          counters={defaultLoadingCounters}
+        >
+          <ContentWrapper>
+            <VSpacer size={8} />
+            <InitiativeTimelineComponent initiativeId={initiativeId} />
+            <VSpacer size={32} />
+          </ContentWrapper>
+        </InitiativeDetailsBaseScreenComponent>
+      ),
+      initiative => (
+        <InitiativeDetailsBaseScreenComponent
+          initiative={initiative}
+          onHeaderDetailsPress={navigateToBeneficiaryDetails}
+          counters={getInitiativeCounters(initiative)}
+          footer={getInitiativeFooterCta(initiative)}
+        >
+          {getInitiativeDetailsContent(initiative)}
+        </InitiativeDetailsBaseScreenComponent>
+      )
     )
-  );
-
-  const renderFooter = () => {
-    if (
-      initiativeData?.initiativeRewardType === InitiativeRewardTypeEnum.DISCOUNT
-    ) {
-      return (
-        <View style={[IOStyles.footer, styles.footer]}>
-          <ButtonSolid
-            label="Autorizza pagamento"
-            accessibilityLabel="Autorizza pagamento"
-            icon="qrCode"
-            onPress={navigateToPaymentAuthorization}
-            fullWidth={true}
-          />
-          <VSpacer size={8} />
-        </View>
-      );
-    }
-  };
-
-  return (
-    <BaseScreenComponent
-      goBack={true}
-      headerBackgroundColor={IOColors["blueIO-50"]}
-      contextualHelp={emptyContextualHelp}
-      customRightIcon={{
-        iconName: "info",
-        onPress: handleBeneficiaryDetailsPress
-      }}
-    >
-      <FocusAwareStatusBar
-        backgroundColor={IOColors["blue-50"]}
-        barStyle={"dark-content"}
-      />
-      <ScrollView
-        contentContainerStyle={styles.scrollContainer}
-        scrollIndicatorInsets={{ right: 1 }}
-      >
-        {cardComponent}
-        {renderScreenContent()}
-      </ScrollView>
-      {renderFooter()}
-    </BaseScreenComponent>
   );
 };
 
+const defaultLoadingCounters: ReadonlyArray<InitiativeBonusCounter> = [
+  {
+    type: "AmountWithProgress",
+    isLoading: true
+  }
+];
+
 const styles = StyleSheet.create({
-  scrollContainer: {
-    flexGrow: 1
-  },
   newInitiativeMessageContainer: {
     alignItems: "center",
     justifyContent: "center",
     padding: 32,
     flex: 1,
     flexGrow: 1
-  },
-  lastUpdate: {
-    alignSelf: "center",
-    alignItems: "center",
-    padding: 16
-  },
-  footer: {
-    paddingBottom: 20
   }
 });
